# docs/swagger.yaml
basePath: /
definitions:
  controller.Info:
    properties:
      message:
        example: 消息
        type: string
      ok:
        example: true
        type: boolean
    type: object
  controller.JSONResult:
    properties:
      code:
        example: 0
        type: integer
      data:
        type: object
      info:
        $ref: '#/definitions/controller.Info'
    type: object
info:
  contact: {}
  description: 网页版kubectl cp
  title: KubeCp Swagger
  version: "1.0"
paths:
...
  /api/k8s/exec:
    get:
      description: 在pod的容器中执行
      parameters:
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: command
        type: array
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: stderr
        type: boolean
      - in: query
        name: stdout
        type: boolean
      - in: query
        name: tty
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: Container 执行器
      tags:
      - Kubernetes
  /api/k8s/file_browser:
    get:
      description: 容器文件浏览器
      parameters:
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: path
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: path
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: path
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: path
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: FileBrowser
      tags:
      - Kubernetes
...
  /api/k8s/namespace:
    get:
      description: 命名空间列表
      parameters:
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: ListNamespace
      tags:
      - Kubernetes
  /api/k8s/pods:
    get:
      description: 命名空间下Pod资源列表
      parameters:
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: pod
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: pod
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: pod
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: pod
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: ListNamespaceAllSource
      tags:
      - Kubernetes
...
swagger: "2.0"